% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddif.R
\name{ddif}
\alias{ddif}
\title{DIF Evaluation}
\usage{
ddif(idata, group, type)
}
\arguments{
\item{idata}{a matrix or data.frame of binary scored item responses. Only
includes item data}

\item{group}{a numeric vector of values identifying group membership.  The
focal group should be coded as 1 while the reference group should be coded
as 0. Missing values are not allowed for \emph{group}.}

\item{type}{identifies the type of method to use.  Options include "mh"
for Mantel Haenszel, "lr" for Logistic Regression, or "all" for both
methods}
}
\description{
\code{ddif} identifies DIF using MH, LR, or both approaches
}
\details{
The \code{ddif} function calls the \code{\link{dlr}} and/or \code{\link{dmh}}
depending on the \emph{type} of method specified.  If the \emph{type} is
"all", the results will also display output of DIF magnitude categories for
method in a data frame.
}
\seealso{
\code{\link{dmh}} for MH DIF method
\code{\link{dlr}} for LR DIF method
\code{\link{R2N}} for Nagelkerke R^2
\code{\link{mag}} for magnitude categorization
}

