% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difdat.R
\docType{data}
\name{difdat}
\alias{difdat}
\title{Sample data to demonstrate how to use functions}
\format{A dataframe with 1500 rows and 21 variables:
\describe{
  \item{G}{group membership, focus(1) and reference(0)}
  \item{I1}{item 1}
  \item{I2}{item 2}
  \item{I3}{item 3}
  \item{I4}{item 4}
  \item{I5}{item 5}
  \item{I6}{item 6}
  \item{I7}{item 7}
  \item{I8}{item 8}
  \item{I9}{item 9}
  \item{I10}{item 10}
  \item{I11}{item 11}
  \item{I12}{item 12}
  \item{I13}{item 13}
  \item{I14}{item 14}
  \item{I15}{item 15}
  \item{I16}{item 16}
  \item{I17}{item 17}
  \item{I18}{item 18}
  \item{I19}{item 19}
  \item{I20}{item 20}
  }}
\usage{
difdat
}
\description{
A dataset containing a variable for group identification and 20 items
The data was simulated for two different groups
}
\keyword{datasets}

