% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr.R
\name{dlr}
\alias{dlr}
\title{Logistic Regression DIF Function}
\usage{
dlr(idata, group, alpha = 0.05)
}
\arguments{
\item{idata}{a matrix or data.frame of binary scored item responses. Only
includes item data}

\item{group}{a numeric vector of values identifying group membership.  The
focal group should be coded as 1 while the reference group should be coded
as 0. Missing values are not allowed for \emph{group}.}

\item{alpha}{identifies the alpha value used in the analyses. The default
value is 0.05.}
}
\value{
A list containing:

LRModels        a list containing three data frames consisting of the
                results from the M0 M1 and M2 LR models

UniformDIF      a data frame consisting of the uniform DIF results

NonuniformDIF   a data frame consisting of the uniform DIF results
}
\description{
\code{dlr} identifies DIF using the logistic regression statistic method
}
\details{
The dlr function computes logistic regression results within a DIF
framework specified by Swaminathan and Rogers (1990).  Three models are
traditionally fitted in order to determine if an item has mixed,
nonuniform, or uniformed DIF.  The three models that are fitted include:

\deqn{M0: logit(\pig) = \beta0 + \beta1\theta + \beta2g + \beta3\thetag}
\deqn{M1: logit(\pig) = \beta0 + \beta1\theta + \beta2g}
\deqn{M2: logit(\pig) = \beta0 + \beta1\theta}

where
\eqn{\pig} is the probability of correctly answering an item in
group \emph{g},
\eqn{\theta} is examinee’s matching criteria, and
\emph{g} is the group membership of the examinee.

Nonuniform DIF is assessed by comparing the \emph{M0} and \emph{M1} models.
Uniform DIF is evaluated by comparing the \emph{M1} and \emph{M2} models.

The magnitude of DIF is calculated by subtracting the \emph{R}^2 estimate
of the compact model from the \emph{R}^2 value of the augmented model.
Nagelkerke  R^2 estimate (Nagelkerke, 1991) is used to compute the DIF
effect size. Classification indices of the magnitude of DIF are based on
values suggested by Zumbo and Thomas (1997).  classifications of \emph{A},
\emph{B}, and \emph{C} represent negligible, moderate, and large DIF effect
sizes.

\bold{A}:  delta{R^2} < or = .13
\bold{B}:  .13 < delta{R^2} < or = .26
\bold{C}:  delta{R^2} > .26
}
\examples{
#Loading data
data(difdat)

#Testing DIF using LR approach alpha = .05
dif.lr <- dlr(difdat[, 1:21], difdat[, 1])

#Testing DIF using LR approach alpha = .01
dif.lr <- dlr(difdat[, 1:21], difdat[, 1], alpha = .01)

}
\references{
Hidalgo, M. D. and Lopez-Pina, J.A. (2004). Differential item functioning
detection and effect size: a comparison between logistic regression and
Mantel-Haenszel procedures. \emph{Educational and Psychological
Measurement}, 64, 903-915.

Nagelkerke, N. J. D. (1991). A note on a general definition of the
coefficient of determination. \emph{Biometrika}, 78, 691-692.

Swaminathan, H. and Rogers, H. J. (1990). Detecting differential item
functioning using logistic regression procedures. \emph{Journal of
Educational Measurement}, 27, 361-370.

Zumbo, B. D. (1999). \emph{A handbook on the theory and methods of
differential item functioning (DIF): Logistic regression modeling as a
unitary framework for binary and Likert-type (ordinal) item scores}. Ottawa,
Canada: Directorate of Human Resources Research and Evaluation, Department
of National Defense.

Zumbo, B. D.,&Thomas, D. R. (1997). \emph{A measure of effect size for a
model-based approach for studying DIF} (Working paper of the Edgeworth
Laboratory for Quantitative Behavioral Science). Prince George, Canada:
University of Northern British Columbia.
}
\seealso{
\code{\link{R2N}} for Nagelkerke R^2
}

